docs: Adds explanation of how to work with other metrics
This commit adds an example to explain how to use metrics which are outside the scope of this library (which is going to be a really common use-case.) The example given is very generalized; demonstrates a single metric that can be used as a foundation to follow-along with the examples provided by the prom-client library itself. We also include a link directly to the examples that will be relevant for newcomers looking for the fastest answer.